Move Didyoumean file into its own package
