+++ src/main/java/seedu/bookmark/algo/Algo.java
@@ -1,0 +1,4 @@
+package seedu.bookmark.algo;
+
+public interface Algo {
+}
