+++ src/main/java/seedu/bookmark/algo/AlgoManager.java
@@ -1,0 +1,4 @@
+package seedu.bookmark.algo;
+
+public class AlgoManager {
+}
