+++ src/main/java/seedu/bookmark/algo/didyoumean/WordStoreComparator.java
@@ -1,4 +1,4 @@
-package seedu.bookmark.algo;
+package seedu.bookmark.algo.didyoumean;
 
 import java.util.Comparator;
